feat: relative path field
