+++ lib/update.ts
@@ -7,6 +7,8 @@ import { semver, colors } from "../deps.ts";
 import { getLatestVersion } from "./registries.ts";
 import { box } from "./utils.ts";
 
+const oneDay = 1000 * 60 * 60 * 24;
+
 export class UpdateNotifier {
   public name = "";
   public alias = "";
@@ -95,5 +97,34 @@ export class UpdateNotifier {
 
   needCheck() {
     return Date.now() - this.lastUpdateCheck > this.updateCheckInterval;
+  }
+}
+
+/** Install update handler cli to check for updates and notify user */
+export async function installUpdateHandler(
+  module: string,
+  executable: string,
+  updateCheckInterval: number = oneDay,
+) {
+  const installation = Deno.run({
+    cmd: [
+      "deno",
+      "install",
+      "--unstable",
+      "-f",
+      "-A",
+      "-n",
+      module,
+      "https://x.nest.land/hatcher@0.8.0/cli.ts",
+      executable,
+      updateCheckInterval.toString(),
+    ],
+  });
+
+  const status = await installation.status();
+  installation.close();
+
+  if (status.success === false || status.code !== 0) {
+    throw new Error("Update handler installation failed.");
   }
 }
